feat(tokens): add typings
