--- packages/tokens/src/global/elevations.js
@@ -1,31 +1,0 @@
-import { elevate } from '@gympass/yoga-common';
-import colors from './colors';
-
-/**
- * @module elevation
- * @desc Elevation (z-index) tokens module.
- *
- * @memberof @gympass/yoga-tokens
- */
-
-/**
- * An elevation
- * @typedef Elevation
- *
- * @type {Object}
- * @property {String} zero
- * @property {String} small
- * @property {String} medium
- * @property {String} large
- */
-
-/**
- * @type Elevation
- * @default
- */
-const elevations = elevate({ color: colors.medium });
-
-[elevations.zero, elevations.small, elevations.medium, elevations.large] =
-  elevations;
-
-export default elevations;
