+++ packages/tokens/src/global/font-sizes.ts
@@ -1,0 +1,27 @@
+export interface IFontSizesProps extends Array<number> {
+  xxsmall?: number;
+  xsmall?: number;
+  small?: number;
+  medium?: number;
+  large?: number;
+  xlarge?: number;
+  xxlarge?: number;
+  xxxlarge?: number;
+  huge?: number;
+}
+
+const fontSizes: IFontSizesProps  = [10, 12, 14, 16, 20, 24, 32, 40, 48];
+
+[
+  fontSizes.xxsmall,
+  fontSizes.xsmall,
+  fontSizes.small,
+  fontSizes.medium,
+  fontSizes.large,
+  fontSizes.xlarge,
+  fontSizes.xxlarge,
+  fontSizes.xxxlarge,
+  fontSizes.huge,
+] = fontSizes;
+
+export default fontSizes;
