+++ packages/tokens/src/global/fonts.ts
@@ -1,0 +1,19 @@
+import weights from './font-weights';
+
+export interface IFontProps {
+  family: string;
+  weight: (number | string)[];
+}
+
+export interface IFontsProps {
+  rubik?: IFontProps;
+}
+
+const fonts: IFontsProps = {
+  rubik: {
+    family: 'Rubik',
+    weight: [...weights, ...weights.map(weight => `${weight}i`)],
+  },
+};
+
+export default fonts;
